feat(wl-id): Allow WL ID to be changed
The uid=1000(andre) gid=1000(andre) grupos=1000(andre),4(adm),27(sudo),123(lpadmin),999(docker) of a  can now be updated via the API.
The  is introduced to handle the renaming of the ID and updating related tables.
An integration test is added to verify the functionality.
Database migrations are added to set  on foreign key constraints.